sys_open() first version (end)
Conclusion (or a sort of a test) of the sys_open() plus the definition of the filetable in the thread.h structure (the structure that calls also the proc structure for all the processes currently present on the disk).
This is only a test, I'm not sure if there are all the things that we need and if it can work but it is a starting point.