First prototype for the sys_fork() system call
There are some comments on the sys_call that explain it all.
In thread.h I commented the filetable line (we defined it in proc.h).
In addrspace.c I specified that the copy of newas and oldas is a one to one copy (we don't have to implement the VM so we don't have page table and so on to copy).

Co-Authored-By: marcochiavello <110553306+marcochiavello@users.noreply.github.com>