Some adjustments on the fork and sys_exit()
Prototype of sys__exit() following the os161-userprocesses pdf loaded by the professor. Starting the waitpid, a doubt occurs: we need to create a process table that takes into account all the proc structures in order to have the possibility to search the structure (and load it) in waitpid only using the pid of the process that we are waiting for.

Co-Authored-By: marcochiavello <110553306+marcochiavello@users.noreply.github.com>